refactor: create zero-session-service and extract zero session ops from agent-session-service
Move all zero agent session table operations into a new zero-layer service,
making agent-session-service pure infra (no zero schema imports). Also fix
secondary layering violations in chat-thread-service and export-service.

Closes #7679
