+++ turbo/apps/web/app/api/agent/sessions/[id]/route.ts
@@ -6,7 +6,7 @@ import {
 import { sessionsByIdContract } from "@vm0/core";
 import { initServices } from "../../../../../src/lib/init-services";
 import { getAuthContext } from "../../../../../src/lib/auth/get-auth-context";
-import { getSessionResponse } from "../../../../../src/lib/agent-session/agent-session-service";
+import { getSessionResponse } from "../../../../../src/lib/zero/zero-session-service";
 import { isNotFound, isForbidden } from "../../../../../src/lib/errors";
 
 const router = tsr.router(sessionsByIdContract, {
