+++ turbo/apps/web/app/api/internal/callbacks/chat/__tests__/route.test.ts
@@ -15,7 +15,7 @@ import {
 } from "../../../../../../src/__tests__/api-test-helpers";
 import { computeHmacSignature } from "../../../../../../src/lib/callback/hmac";
 import { reloadEnv } from "../../../../../../src/env";
-import { getSessionChatMessages } from "../../../../../../src/lib/agent-session";
+import { getSessionChatMessages } from "../../../../../../src/lib/zero/zero-session-service";
 import { POST as createThreadHandler } from "../../../../zero/chat-threads/route";
 import { POST } from "../route";
 import { http } from "../../../../../../src/__tests__/msw";
