+++ turbo/apps/web/src/lib/agent-session/index.ts
@@ -2,7 +2,4 @@ export {
   getAgentSessionWithConversation,
   createAgentSession,
   updateAgentSession,
-  listAgentSessions,
-  appendChatMessages,
-  getSessionChatMessages,
 } from "./agent-session-service";
